Split JSON building from event finding logic
