+++ src/JsonBuilder.php
@@ -4,8 +4,6 @@ declare( strict_types = 1 );
 
 namespace ModernTimeline;
 
-use ModernTimeline\ResultFacade\PropertyValueCollection;
-use ModernTimeline\ResultFacade\Subject;
 use ModernTimeline\ResultFacade\SubjectCollection;
 use ModernTimeline\SlidePresenter\SlidePresenter;
 use SMWDITime;
@@ -19,40 +17,7 @@ class JsonBuilder {
 	}
 
 	public function buildTimelineJson( SubjectCollection $pages ): array {
-		$events = [];
-
-		foreach ( $pages->getSubjects() as $subject ) {
-			[ $startDate, $endDate ] = $this->getDates( $subject );
-
-			if ( $startDate !== null ) {
-				$events[] = new Event( $subject, $startDate, $endDate );
-			}
-		}
-
-		return $this->eventsToTimelineJson( $events );
-	}
-
-	private function getDates( Subject $subject ): array {
-		$startDate = null;
-		$endDate = null;
-
-		foreach ( $this->getPropertyValueCollectionsWithDates( $subject ) as $propertyValues ) {
-			$dataItem = $propertyValues->getDataItems()[0];
-
-			if ( $dataItem instanceof SMWDITime ) {
-				if ( $startDate === null ) {
-					$startDate = $dataItem;
-				}
-				else if ( $endDate === null ) {
-					$endDate = $dataItem;
-				}
-				else {
-					break;
-				}
-			}
-		}
-
-		return [ $startDate, $endDate ];
+		return $this->eventsToTimelineJson( ( new EventExtractor() )->extractEvents( $pages ) );
 	}
 
 	/**
@@ -83,19 +48,6 @@ class JsonBuilder {
 		}
 
 		return $jsonEvent;
-	}
-
-	/**
-	 * @return PropertyValueCollection[]
-	 */
-	private function getPropertyValueCollectionsWithDates( Subject $subject ) {
-		return array_filter(
-			$subject->getPropertyValueCollections(),
-			function( PropertyValueCollection $pvc ) {
-				return $pvc->getPrintRequest()->getTypeID() === '_dat'
-					&& $pvc->getDataItems() !== [];
-			}
-		);
 	}
 
 	private function newHeadline( \Title $title ): string {
